Improve compiler flag logic
If FFTW has been built with the combined_threads option it will not
require linker flags therefore we can't use flags to work out what
threading it uses internally. Therefore if the rest of the program
is using OpenMP we will set the compiler flag -DHAVE_FFTW3F_THREADS
so that the fftw threading callback is replaced. This should make
no difference if it's already using OpenMP but will prevent thread
conflict if it's using pthreads.