Enforce correct wisdom file at time of use
This commit moves the call to set_wisdom_file from when initfile is
read to just before fftw is used. This means it appears in more
places in the code but it is more foolproof in terms of ensuring
that the correct wisdom is loaded, for example if a user changes the
fftw threading option.