Initial support for StarNet 2T
Adds support for starnet2T. The initial Starnet2T release is for MacOS
M1/M2 however the Siril interface is generic so as long as the command
name follows the expected pattern, support should be automatic.

Currently the revised input and output args are supported
(-i input.tif instead of just input.tif and -o output.tif similarly).

The -s argument is supported to provide a custom stride arg.

The -u argument is supported to enable upscaling. For starnet2T Siril
will not carry out upscaling / downscaling but will leave it to
StarNet.

The input mask argument is *not* currently supported as Siril currently
does not have anything to provide for a mask. However application-wide
mask support is something that has been discussed as an area of interest
for a future Siril development cycle and it is expected that the -m
argument would be added alongside introduction of masks in Siril
generally.