Fix #1034
The image preview update functions were causing the problem, as they
still applied even when the dialog was closing and after histo_close
was called. Fixed using the static gboolean `closing` which is set
TRUE on opening the dialog and FALSE at the start of the close callback,
and is used to prevent the preview update callback if the dialog is
closing. This bug had always been present but was highlighted by
a log message added recently in copyfits()