Suppress percentage outputs in Siril log for Torch StarNet versions
This commit sets doprint to false when it detects "Working:" in the
output, which is necessary to parse the new Torch versions correctly.
After this we don't want to print any more StarNet output in the
Siril log, we just want to update the status bar and check for
completion.