Static analysis fixes in starnet.c
1. Initialize workingfit and fit to { 0 }
This is needed because if the executable check fails the code jumps
to CLEANUP without otherwise initialising these variables, after
which clearfits() is called on them which would not go well.

2. At cleanup, check currentdir != NULL before calling g_chdir() on
it. Again this is due to the possibility of the executable check
failing before the current directory is assigned to currentdir.