Implemented single GUI gnuplot handle gnuplot_init() now takes a gboolean to indicate if it is being called to initialize gui.gplot or a non-GUI gnuplot handle
Fixed hang in sequence working by limiting to 1 thread
This is still quick and given a potential change of plotting
backend in 1.3 is all the effort I intend to spend here.

I've used the wxt gnuplot term type for graphical plots, need to
check if this is available on all platforms.

Also added the possibility of reusing an existing plot by setting the
handle->reuse gboolean.

The handle now copes with unlimited number of plot windows, and stores the
number of plots separately for each window.