Close liveview dialog during undo/Redo
It is quite an issue to have hit undo/redo when a processing tool dialog is opened.
Especially if this one is a liveview one and if we undo a crop process.
The best thing to do, at least for the next release, is to close every liveview
dialogs when hitting undo/redo.