Revert the custom button + filechooser experiment
Revert to using GtkFileChooserButton in line with the rest of the
code, for consistency prior to a GTK4 migration push. I will
make a diff of the change and add it to the MR discussion, so we
have a record of it in case it is of use when doing the migration.