Finish SCNR optimization, check in WIP
Optimizing SCNR has given me some more ideas on how to further
optimize other transforms. In particular, timings show it is faster
using OMP to parallelize calls to cmsDoTransform than it is using
the threaded plugin. The display routine already does this if
the "optimising" use_extra_mem parameter is off, so I'll do some
more timings but it looks like it will actually be better for
performance as well as memory to get rid of that option, and the
threaded plugin altogether, and use fast_float to benefit from SIMD
along rows and OMP to parallelize the rows themselves.