Update photometry unit test
Each time we update PSF.c, these tests fail because of slight changes
in values to an accuracy which is far from what we search for.
The tests have been modified to loosen a bit the accuracy required.
Changes are also made to the display format so that it is easier
to spot the errors (basically, the values are printed against ref and
accuracy).

Also small editorial changes in the comments from last commit