Remove unused function, improve display transform efficiency
On profile change etc. a check is done if the display profile is
identical to the gfit profile. If it is, the display transform
is skipped even if it would otherwise be required, as it would
be a null transform.
Remove an unused function.
Make refresh_icc_transforms only run when required rather than in
every call to notify_gfit_modified().