Fix color management in starnet
Starnet does not do color management, so the returned tiff has
no color profile and is automatically assigned one on loading which
is probably wrong. So we make a copy of the original color profile
and reapply it to both the starless and star mask image after they
are loaded back in.