Add hinting for previously stretched files, bugfix
* Add hinting for previously stretched files. The file open dialog
now contains a checkbox "Stretched FITS" with tooltip. If the file
to be opened has been stretched in a previous version of Siril that
does not annotate stretches in the FITS HISTORY header, this
checkbox can be used to hint that the sRGB ICC profile should be
assigned. It has no effect if the FITS file does contain an ICC
profile, or for any other file format.
* The `load` command now accepts an optional argument
`-stretch_hint`, which is handled in the same way as the checkbox.

Bugfix: JPEGs etc. were being assigned a V2 profile instead of V4
when loaded. This is now fixed.