Allow scrolling of the star recomposition UI
Useful for if the advanced UI is selected, which can be too big
for some users to fit on screen.
This requires GtkScrolledWindow propagate natural width, which was
introduced in GTK 3.22 so requires a bump to the minimum required
version of GTK. 3.22 is itself ancient so hopefully this isn't a
problem.