A different approach to showing the unmerged changes.
This now correctly shows all unmerged changes by starting at the
FETCH_HEAD and checking whether the commit oid is the same as the
HEAD oid. If it is, we need go no further; if it isn't, this is an
unmerged commit and we show the commit message.

Currently I'm using the siril_confirmation_dialog() which looks
like it will get messy if there are lots of unmerged commits with
long commit messages, so I'll probably have to make a different
function that provides a similar confirmation dialog but with the
secondary text inside a GtkScrollableWindow.