Revert change to Open dialog
Change introduced hinting for previously stretched files. This is
no longer done, though it could be reintroduced via
icc_auto_assign_and_convert. That would probably be a good idea,
it would work for a substantial number of previously-stretched
files and when it doesn't we're no worse off, the user just has to
re-assign or convert the color space themselves.