Remove some dead code from starnet.c
This code set up a starmask filename and called starnet_torch with
the -m starmask.tif argument, but did nothing with this
starmask except immediately delete it. The actual star mask is
produced in all cases in exactly the same way, using imoper(), so
this code was just wasting starnet CPU time and a bit of temp
disk space.