Updated Deconvolution dialog to use a (manual) preview
Use of copy_gfit_to_backup() et al, plus a hidden "Preview in ROI"
button, allows for much closer behaviour to the image operations
with live previews. If the ROI is moved, a recalculation will
update the ROI; if Apply is clicked while a ROI preview is in place
then the effect won't get double-applied to the ROI and the dialog
behaves more predictably when it is closed without applying.