Improved ROI preview added for denoise, fixed geometry ops bugs
Added ROI preview for denoise similar to deconvolution
Fixed bugs where doing geometry ops that could change the image
size with a ROI set could result in the ROI extending past the
dimensions of the new image and crashing. Even where this is not
the case, retaining a ROI after a geometry op gives undesirable
results, so the policy is simple that a ROI will be removed on
any geometry op.