Replace interpolation functions
I realised we already had a between-4-pixels interpolation function
in the deconvolution helper functions file, so I've replaced my
function with this one and a variation of it that handles WORD data.
This has fixed the issue with dips in the middle of what should be
a smooth profile through e.g. a star.