Altered channel split behaviour
Channel split now leaves all component values untouched so FITS
are produced with values in the color space's natural range
(e.g. CIE L* in [0..100], a* in [-127..128] etc.)

The FILTER card is marked accordingly (e.g. 'LAB L' etc.)

The components are re-scaled to Siril ranges when loading a file
with a recognised FILTER card that requires re-scaling, and are
scaled back to the proper values again when saving.

This allows processing of extracted channels in Siril's standard
fp range of [0..1] but correct values for the colorspace when
read using other software.

TODO: haven't done USHORT conversions yet. Need to understand the
standard method for representing e.g. a* and b* ranges in USHORT
data elements.