JpegXL color profiles work.
However the loaded JpegXL has the color profile of the internal
representation, not the original data. It is therefore necessary
to convert back to the desired ICC profile.
The JpegXL discord suggests returning an image in the original
colorspace is still a work in progress, so this may be the best
that can be done for now.