Fix #1226
Altering the image color profile clears the ROI.
The amount of possibilities around images being in preview, undo
states etc. make this more complex than it's worth. I might be
able to make it work, but I think it would always be a fragile
bit of code. It seems unlikely that having an ICC profile change
reset the ROI will be disruptive to anyone's workflow.