Add color management to photometric color calibration
Ensures the color temperature is correct in the user's working
color space.
Prints a warning if the image has a nonlinear color profile.
If no color profile is attached, it will assign a linearised
version of the user's preferred RGB color space and use that.

This represents a behaviour change and will therefore require
a companion documentation MR.