JPEG XL fixes
JPEG XL is mostly behaving. Read and write of color and mono images
works.

Sometimes the name of the ICC profile is different when a saved
image is loaded back in (changed to a standard format of some kind,
apparently by libjxl, and I'm not yet sure why). Functionally
everything is the same, but the name is different which could be
confusing.