Fix chromatic adaptation
The colours obtained now look visually almost identical to master
and very similar to Cyril's PI result: the knee point in the curve
obtained by plotting the debug data occurs at B-V = 0.47.

I've also added a second function for conversion of TempK to xyY,
this one was written by BQ Octantis and explained at
https://www.cloudynights.com/topic/849382-generating-a-planckian-ccm/
It doesn't claim to extend down to as low temperatures as the Kim et
al spline function and I think it's slower to compute because of the
calls to pow(), but it's useful to gain confidence in the robustness
of the method. You can see the difference by commenting out either
temp_to_xyY() or BQ_temp_to_xyY() at lines 115-116 in
algos/photometric_cc.c