Add additional temp-to-xyY converter
This is based on Mitchell Charity's work on black body
temperatures. It uses his tabulated data as a LUT and uses
linear interpolation between the values. It is claimed to be
valid down to 1000K.

Again, this one is just added for comparison against Kim's 3rd
order splines and BQ Octantis's 6th order splines. Agreement is
very good.