pass a more comprehensive table to asnet
When solving with distorsions, asnet can produce inaccurate solutions
if we pass only stars x,y columns.
We improve the solution by passing also flux and bkg (as A and B from
star_finder) and letting asnet resort and do whatever it sees fit with that
info...
In some cases still, it may not honor the --crpix-center flag. We recompute it
and pass it to fit->wcsdata so that the reporting is correct.