SPCC now uses the repository
SPCC now reads sensor and filter data from the local repository.
The filter widgets have been split into R, G and B combos and
the combos auto-populate from the available objects in the SPCC
database.

TODO: add a channel field to the schema for MONO_FILTER objects
TODO: make the handling of OSC sensors more robust (it's currently
reliant on the order of objects in the JSON, would be better if
there was a channel number for each channel response).