Fix SPCC maths (draft)
Draft because I still want to understand this a little better
but it gives much better white balance control and the process
works for the DSLR test image.
Also need to decide betwee the pure GSL linear fit function and
the repeated medians one. Which is faster, is the faster one
good enough?