Add command interface for SPCC
The interface extends the process_pcc function.
Sensors and filters can be specified by name as arguments: this
may be cumbersome, however if the sensor / filter arguments are
omitted then the values in com.pref will be used, so a non-GUI
alternative is to use the set command to set the user's preferred
sensor and filters, after which spcc only requires the
'-osc={true|false} argument to specify whether to use mono or
OSC preferences.