Fix for map_image_coordinates_wcs
Untested yet but I had failed to allocate a status array; fixing
this should allow it to work.

Also I've updated the function to work a row at a time, in order
to strike a better balance between allowing wcslib to work on
vectors and not requiring vast amounts of memory.