Initial code to allow use of master flat / master bias as weight
This allows the (bias subtracted if necessary) master flat set in
the preprocessing tab to be used as an initial pixel weighting for
all frames. Thus the weight of pixels between frames will take into
account the position relative to the master flat of each input
pixel value.

This is currently enabled / disabled by hard coded booleans, it
still needs UI elements adding. Currently disabled by default until
I find a suitable sequence and calibration frames to test it with,