Proper fix for deconvolution and denoise undo / redo issues
Fixes #1276 and #1279

These are now correctly treated as dialogs with preview
functionality and will close when undo is called. This results in
removal of the hidden preview state that is necessary between ROI
operations and final application but ensures that when applying
after undo / redo the process is always working on fresh data.