Remove old stacking upscaling and the old register -drizzle options
The register -drizzle option is removed as the true drizzle command
seqapplydrizzle is now preferred.
The upscale-at-stacking functionality has been removed for
simplificaton and consistency in that upscaling, whether by drizzle
or other methods, is now always applied before stacking.

Simple upscaling can be achieved using options to the seqextract_Ha
and seqextract_HaOIII commands or using the soon-to-be-added
sequpscale command. But in general these are only useful for when
you have very few subs - with 10 or more (for x2 upscaling) true
drizzle should be preferred.