SPCC data bugfixes
Since allowing objects to be added to multiple lists Siril could
crash if the lists were freed and reloaded.
This became apparent when making the lists refresh on manual
update (which required fixing an incorrect widget name).
This has been fixed by checking if an object has already been
added to a list, and if so a copy is made before adding to any
additional lists.