Add background / gradient subtraction for spectroscopic profiles
With spectroscopic data and tri-cut selected, cut uses the outer
profile lines as background strips and subtracts a background
value (allowing for a linear gradient) along the central strip.
The background value is computed using the robust linear fit.

Added following discussions with Robin Leadbeater on CN and based
on his presentation on reducing spectra using ISIS. More work is
needed to make Siril a first rate application for spectral
reduction but this is another step on the road and a substantial
improvement to the initial implementation.