Fix memleak in ROI processing
When calling copyfits with CP_FORMAT, no check was being done to
see if there were data structures that needed to be freed, so
allocated pointers could be set to NULL and subsequently realloced
resulting in a memory leak. I've added a call to clearfits() at
the start of the CP_FORMAT section to protect against this.