UI improvements
UI elements only show for applicable methods, in particular some
widgets only applicable to global registration have been moved to
the notebook and the output registration widgets are hidden for
non-output methods.

There's still a bit too much empty space for some methods, this is
due to the GtkStack size being fixed to the vertical space of the
largest page.