Complete implementation of Gaia status widget
The status is automatically checked when the SPCC dialog is opened,
it can be updated by clicking the button and the status is
reported by means of a log message and an icon within the button.
The colours and status messages match the traffic light colours on
the Gaia archive status webpage.