Fix bug in gaia archive check idle, improve fetch_url behaviours
Use the verbose flag in fetch_url_async_data - if not verbose, no
output on failure will occur.

Add abort_on_fail flag to fetch_url and an element in
fetch_url_async_data. If this is TRUE then only 1 attempt to
fetch the data will occur. This is used in the Gaia status check,
where we want a quick indication.