Add GraXpert ROI processing
This works for denoising, and is not allowed for background
extraction (it doesn't make sense doing background extraction on
a subset of the image). It isn't quite as good as ROI processing
for other processing-heavy functions as there is a significant
startup time for the AI model, but it helps (on my test image
processing a ROI takes ~54s as opposed to a couple of minutes to
process the entire image).