Fix issue with conversion of .avi sequences
For a film like b.avi, accepting the confirmation dialog choice to
convert it would result in a film_file->filename of b.avi;
convert_single_film_to_ser would then produce a destroot of
b.avi_converted; and this would subsequently be stripped back to
a sequence name of b to open, which would result in Siril trying to
open b.avi again and re-showing the confirmation dialog.

This change introduces a slight alteration in the setup of destroot
so that the existing extension is removed, resulting in a destroot
of b_converted and successfully opening the converted sequence on
completion.