Fix #1346
Adds some log messages highlighting that the saved and re-read
files are temporary ones.

Fix a bad assumption in open_graxpert_result that the working fits*
is always &gfit, actually it should be args->fit

Instwad of copying the entire result fits to args->fit, only move
data and fdata (this fixes the part of the issue where the header
could no longer be read after performing 2 consecutive graxpert
operations).