Improve the code by using a new function fits_swap_image_data()
Based on the C++ swap method concept. In the case of fits data it
allows swapping the image data from a temporary fits structure into
a wanted one, and then the old data from the wanted one is freed
when clearfits is called on the temporary one.

Also make the Graxpert window non-resizeable.