Minor polish for the curves MR
Fix a couple of potential null dereferences pointed out by coverity

Make the vertical histogram zoom always 1.0 (the histogram stretch
code is a bit confusing here and references a GtkAdjustment but
that is always 1.0, it should probably be tidied up at some point).

Make a couple of variables static as they are not used outside the
file.