Add `pwd` command to print the current working directory
Not really useful for scripting as the script log will show output
generated by the `cd` command, but it's occasioanlly useful when
dealing with a complex project if you lose track of which directory
you're in. It does say at the top but in an ellipsized form which
is no good for copying & pasting into a file explorer, say.