Backport macOS build pipeline
The newer build pipeline significantly simplifies the update
procedure. Introduced in master for the upcoming Siril 1.4, this was
originally not planned to land in 1.2 as it raises system requirements
from macOS High Sierra 10.13 to macOS Catalina 10.15. However, we're
updating 1.2 a lot longer than expected, so I need to make my life
easier now by moving to the newer build pipeline.