Add a preference for the maximum slice size
By setting this to e.g. 2048, you can guarantee that it will be
used for most sliced image operations thus allowing the fastest
possible FFT performance. (For this reason the options are limited
to powers of 2, which are guaranteed to be the fastest anyway and
should require minimal planning to achieve the best performance.)