Add siril.gui_block() and siril.gui_unblock() to the module
These are available for use by python scripts, however the default
remains for the widgets to be blocked. If a script wishes to
unblock them it needs to call siril.gui_unblock(). This could be
useful for helper scripts that support the user while they're still
using the GUI, but should be used with caution in any script that
relies on the Siril state remaining as it was when the script
started.