Fix crash when running certain scripts repeatedly
The issue was that python scripts weren't setting com.script, which
resulted in some GTK operations being executed directly rather than
in an idle in the GTK thread.

Python scripts now set com.script while running, and the redraw()
function now bails on com.headless instead of com.script - this
change needs to be checked for unintended consequences, we might
need to have a com.pythonscript as well that is used for the check
in this case.