Added siril.pipinstall
If siril is started directly, it inherits the system python
environment which will cause problems. However if the venv is
entered at the console using source /path/to/venv/bin/activate and
then siril is started, everything works correctly. This could be
done using a wrapper script but it's an ugly solution, so we need
to find a better way of activating the venv programmatically from
within Siril.