Introduced gui_function()
This is a more robust approach to deciding whether to run GTK
functions directly or using an idle, and accounts for whether to
run them and wait for them (as in a script) or just run them in
an idle in the GTK context.

All functions that are called dependent on a check of com.script
have been updated to the new signature required for use with
gui_function(), i.e. gboolean func(gpointer)

Python scripts no longer automatically set com.script to TRUE, but
they do still set com.python_script