Major updates to setup code, add siril.import_or_install()
The setup code now correctly sets up the venv and checks to see if
there is an installed version of the siril module. If there isn't,
it installs one; if there is, it checks if it's older than the
available one in the system Siril data directory. If the module
needs updating, it is updated. In both cases, if dependencies are
needed they are pulled in.

This commit also adds the siril.import_or_install() method, which
can be used by script writers in place of import: if the required
module is not availble to import, an attempt will be made to
install it and then the import will be re-attempted. This will be
useful to auto-install dependencies without requiring user
intervention. The syntax is:

np = siril.import_or_install("numpy") instead of import numpy as np