Clean up the python API a bit
siril.close_connection() renamed to siril.diconnect()
Removed getters for ImgStats, it is a dataclass so all the
properties are accessible anyway.
Removed getters for FKeywords, it is a dataclass as above. I have
left the setters though, as these provide length bounding for
strings that can be no longer than FLEN_VALUE, and keeping them
for all members provides a consistent interface.