Fix epf and unpurple filters w.r.t threading, previews and ROIs
Both filters need to run in a thread, and both have previews.
When in preview they are controlled by update_preview which handles
the ROI mutex, so the *_filter() function must be passed to
start_in_new_thread().
When doing "Apply" or running the command, we cannot rely on
update_preview to control the ROI mutex so the *_handler() function
must be passed to start_in_new_thread().
This means the ROI mutex is controlled in all cases, the filters
always run in a separate thread and we alway avoid lag in the main
thread.