Revert changesthat broke windows stream monitoring, also retries
Retries mechanism wasn't the right fix for what it was introduced
for.

Windows stream monitoring was a bad change - intended to ensure
that windows stdout / stderr streams were unbuffered but there was
a mistake somewhere so it's reverted. Once tested on Windows we can
see if there's an issue that still needs fixing or not.