Support CFA images in background removal
This is optimal prior to drizzling a CFA sequence. The frames are
treated as independent spatially interleaved subchannels, each of
which is independently background removed, and they are then
recombined.

For sequences that will be drizzled it is optimal because otherwise
you have to do per-frame linear background removal on the drizzled,
pre-stacking frames which (a) are RGB so are 3x the size, and (b)
are an intermediate product that isn't really intended for doing
calculations on (there may be Moiré patterns and varying pixel
contributions).