Add siril.check_module_version("version specifier")
The version specifier can be any pip version specifier (==, <, >=
etc. and can be compound e.g. ">=1.2,<3.4"

This is used to check the current running module meets the script's
requirements and it raises an exception if not. It is also parsed
for when building the list of available git repository scripts in
order to hide those that are incompatible with the module version
detected at startup by the current running Siril instance.