Fix stellar AI model issues
Temporarily remove the check for the AI models for stellar
deconvolution.
This can be reintroduced when fixed upstream, until then stellar
will have to use the latest version. Note this causes a warning
about an unused variable but this will be used again when we
reinstate the check, so I haven't done anything about this.
If the final release of 3.1.0 includes the bug then we may need to
have a little blacklist handler so that 3.1.0 is forced to default
but we allow version checking for later versions once it's fixed.

Also fixes an issue with the AI version check in the command line
that had the effect of making an argument mandatory.