Update docker recipes
libhealpix_cxx will need adding to the MacOS CI, and also we need
to figure out how to add it to the Windows CI as there doesn't
appear to be a mingw-x64 package. Alternatively possibly we could
make it a subproject but unfortunately it's hosted on sourceforge,
not a git repository.