Hidding deconv tool from GraXpert
It has been decided to disable the deconvolution part of GraXpert.
Why is this?
The developer of the deconvolution tool is no longer collaborating with GraXpert.
It is therefore uncertain whether the next version of GraXpert will include the
deconvolution tool. It therefore seems risky to release the tool with version 1.4.
What's more, the new scripting mechanism makes it possible to create an
equivalent tool, and it looks like this is what's going to happen. It will be
easier to maintain.