Update macOS dependencies
This is a big one. GTK is now on 3.24.48 and Cairo is on 1.18.2.
The complete font-related stack has been updated: Pango, HarfBuzz,
FontConfig, with the goal of getting rid of a long-standing bug
(garbled fonts) on macOS. Also updates to cfitsio and LibRaw.

Executables have all been moved to Contents/MacOS. This is where
they belong; having them in Contents/Resources/bin is a (historical)
mistake. Adjust PATH accordingly.