Make all GraXpert tests synchronous, paralleled where possible
The version check runs first then all the AI model checks are done
synchronously in parallel. This slightly improves the time until
the graxpert versions are populated in the combo boxes, and it
avoids having to track child processes for these calls which avoids
edge cases where they might be interrupted using the stop button.