Fix crashes in RGB composition and simplify the dialog
Fix some interactions between the free_generic_seq_args function
introduced in the threadleaks MR and the way 2-step registration
is done in RGB composition. This also rectifies a few possible-
but-not-yet-encountered double frees in a couple of other reg
methods.

This involves adding a gboolean free_sequence parameter to
free_generic_seq_args.

Remove the color calibration widgets from the RGB composition
dialog. Since splitting the platesolve and CC dialogs it doesn't
make sense to have it here any more, and its removal helps to
simplify a busy dialog. CC can still easily be done on completion
of RGB composition, in the usual way.