Change macOS CI for new ephemeral arm64 runner
Use the default work directory /Users/Shared/work so we can utilize the
pre-built bootstrapped JHB artifact to speed up our builds. This wasn't
possible an a shared persistent machine without enforcing a build
concurrency of 1, which is undesireable.

Add the cache directive to preserve ccache.

Update dependencies to ensure a fresh build. This includes only a
jhb update.