Accommodate TF32 use in recent NVidia GPUs in provider test
This prevents the provider failing on the grounds of poor accuracy
between GPU and CPU.

This doesn't mean TF32 is necessarily a good idea for astrophotog
machine learning models. In most image processing it is considered
fine and visually indistinguishable from IEEE754 FP32 use, but in
astrophotography any differences may be magnified during stretching:
we ideally need to find someone with a 30[789]0 or 40?0 GPU who is
willing to do some tests with FP32 enabled and disabled, and see
how much difference we get.