Makes reading astrometric solutions more robust
Some solvers store unit CDELT and PC as CD.
When reading those solutions and using them for astrometric registration, the
CDELTs do not conform the expected convention (scale in deg/px, neg on X) which is an
underlying assumption for computing framing.
Change is introduced to decompose CD as PC/CDELT (in case PC/CDELT formalism is found in the header) to make sure the convention is complied with.
Fixes #1770