Mirrors are now definable using json
The json file ins kept in the notifications folder of master,
same as for the notifications json file.

Note that the primary remote catalogue (in Zenodo) is hard coded
(in case of a problem reading the json file at startup) but must
also be kept as the first element of the json file, which replaces
the initial hard-coding.